split signer and staker
